Move api_v1.py to a directory.
