+++ mymodule/blueprints/api_v1/__init__.py
@@ -1,20 +1,12 @@
-import hashlib
 import flask
 from flask import current_app as app
 from werkzeug import exceptions
 
+from mymodule.blueprints.api_v1 import upload
+
 
 blueprint = flask.Blueprint('v1', __name__, url_prefix='/api/v1')
-
-
-class OutputFile:
-    def __init__(self):
-        self.size = 0
-        self.checksum = hashlib.sha256()
-
-    def write(self, buf):
-        self.size += len(buf)
-        self.checksum.update(buf)
+blueprint.add_url_rule('/upload', view_func=upload.post, methods=['POST'])
 
 
 @blueprint.route('/test')
@@ -38,27 +30,6 @@ def keys():
 def config():
     ns = app.config.get_namespace('MYMODULE_')
     return flask.jsonify(ns)
-
-
-@blueprint.route('/upload', methods=['POST'])
-def upload():
-    results = []
-    storages = flask.request.files.getlist('files')
-    for storage in storages:
-        output = OutputFile()
-        storage.save(output)
-        result = {
-          'size': output.size,
-          'filename': storage.filename,
-          'sha256': output.checksum.hexdigest()
-        }
-        results.append(result)
-        app.logger.info(
-          'filename="%(filename)s" '
-          'size=%(size)d '
-          'sha256="%(sha256)s"' % result
-        )
-    return flask.jsonify(files=results)
 
 
 @blueprint.route('/crud')
